hack: efi: disable reset service
The UEFI reset service hangs on reboot and power off on the CRD and
X13s. Disable it so that we fall back to PSCI.

This is needed until the boot firmware provides an RT_PROP table that
describes the supported services. An alternative is to disable all
runtime services by booting with 'efi=noruntime'.

Not-signed-off-by: Johan Hovold <johan+linaro@kernel.org>