clk: qcom: gcc-sc8280xp: disable pcie3a in clock-controller
The hardware has pcie3 disabled which leads to WARN messages in
dmesg, adapted the code in drivers/clk/qcom/gcc-sc8280xp.c to leave
pcie3 out of the probe. It interacts with the device tree definition

&gcc{
	//this PCIe bus is apparrently disabled in the WDK2023,
	//variable to parse in gcc_sc8280xp_probe()
	pcie_3a_disabled = <1>; // Set to 1 to disable, 0 to enable
};

The code will choose a reduced probe struct if this variable is set.

fixup: added missing additional gdscs that got defined recently

Signed-off-by: Jens Glathe <jens.glathe@oldschoolsolutions.biz>