thermal: qcom: tsens: remove dependency to NVMEM_QCOM_QFPROM
The dependency to NVMEM_QCOM_QFPROM seems to be a problem when using
"make menuconfig", it will exclude QCOM_TSENS from the build wWith
bad effects on power management, observed on the Windows Dev Kit 2023,
and Thinkpad X13s. Removing it doesn't have a negative side effect,
as it seems. Observed this since v6.6.

Signed-off-by: Jens Glathe <jens.glathe@oldschoolsolutions.biz>