arm64: dts: qcom: sc8280xp-wdk2023: dt definition for WDK2023
Device tree for the Windows Dev Kit 2023. This work
is based on the initial work of Merck Hung <merckhung@inblenix.com>.

This dt is more or less deducted from the Thinkpad X13s.
It contains a lot of guesswork, and also a lot of research on
what works with the Windows Dev Kit.

Signed-off-by: Jens Glathe <jens.glathe@oldschoolsolutions.biz>