laptop_defconfig: Add new laptop_defconfig based on Debian
This is completely redone with most changes from Debian pulled in.
Follow the Johan commit to get the list of modules if you don't have it
already.  Additionally, I haven't double checked that the things that
Fedora likes are set properly, and haven't tested webcam either, but the
options should be enabled.