arm64: dts: qcom: sc8280xp: amend definition for arm-smmuv3, set iommu-map
The arm-smmuv3 is responsible for PCIe DMA and usually under control of the
HYP firmware. But if it isn't, when the kernel is booted into EL2,
it needs to be controlled.
The pcie_smmu definition is amended with all available interrupts, and
the msi-parent definition is added. The iommu-map was a mixture of research
with Nikita Travkin <nikita@trvn.ru> in the ACPI dump and lots of tries.
All these settings (also the wired interrupts) have been tested with
a Windows Dev Kit 2023 booted into EL2 with kernel 6.8.

Signed-off-by: Jens Glathe <jglathe@oldschoolsolutions.biz>