dt-bindings: pinctrl: qcom,pmic-gpio: drop pm8008
The binding for PM8008 is being reworked so that internal details like
interrupts and register offsets are no longer described. This
specifically also involves dropping the gpio child node and its
compatible string which is no longer needed.

Note that there are currently no users of the upstream binding and
driver.

Signed-off-by: Johan Hovold <johan+linaro@kernel.org>