UBUNTU: [packaging] remove the kernel configuration for perf
This was introduced by commit "UBUNTU: [Debian] perf -- build in the
context of the full generated local headers". [1]

If you intend to build perf with the latest UAPI headers, you need
to run 'make headers_install'.

Since tools/perf is a userspace program, the kernel configuration
or 'make prepare' should not be necessary.

Remove this, as it never worked as intended.

[1]: https://git.launchpad.net/~ubuntu-kernel/ubuntu/+source/linux/+git/xenial/commit/?id=5531fa5fadf52af8cd0cfcd6ac65a2663cf3a863

Signed-off-by: Masahiro Yamada <masahiro.yamada@canonical.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>