arm64: dts: qcom: x1e80100-pmics: Add ADC
The QCOM SPMI PMIC5 Gen3 ADC in PMK8550 aggregates the ADC readings of all
the PMICs. Most importantly, this is used to read the temperature of
external thermistors, e.g. to obtain the skin temperature of the device.

The placement of temperature sensors is device-specific, so these channels
need to be defined in the board-specific part of the device tree.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>