UBUNTU: [Config] riscv64: Enable Svpbmt support
The page-based memory types is the proper RISC-V extension needed to
support non-coherent peripheral DMAs. Previously this has been done in
two different non-standard ways for the Allwinner D1 and StarFive JH7100
SoCs, but now the proper extension is implemented in the SpacemiT K1 and
hopefully in all future non-coherent SoCs.

Signed-off-by: Emil Renner Berthing <emil.renner.berthing@canonical.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>