UBUNTU: [packaging] avoid piping control-create for error detection
Currently, it is impossible to detect any error in the control-create
script because its output is piped to the sed command.

To address this, separate control-create and sed into distinct commands,
and enable the -e option in the control-create.

With this change, we can prevent situations where we proceed with
invalid debian/control without noticing potential issues in the
control-create.

Signed-off-by: Masahiro Yamada <masahiro.yamada@canonical.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>