UBUNTU: [packaging] remove DESC and =HUMAN= substitution
The sign/unsigned distortion in debian/control has been fixed, but the
Description fields still depend on the build machine's architecture, as
DESC and =HUMAN= are replaced with the build machine's information.

If you generate debian/control on an x86 machine, the unsigned linux-image
entry will look like this:

    Package: linux-image-6.14.0-2-generic
    Build-Profiles: <!stage1>
    Architecture: armhf ppc64el
         [ snip ]
    Description: Linux kernel image for version 6.14.0 on 64 bit x86 SMP
     This package contains the Linux kernel image for version 6.14.0 on
     64 bit x86 SMP.
     .
     Supports Generic processors.
     .

Here, the Architecture field contains "armhf ppc64el", while the
Description field mentions "64 bit x86" because the definition of
'human_arch' for the build arch is included by this line:

    -include $(DEBIAN)/rules.d/$(arch).mk

This commit removes arch-specific information from Description fields.

In Debian kernel, every linux-image entry has arch-specific description,
but that is because the flavour names are arch-specific, such as "amd64",
"cloud-amd64", "arm64", "cloud-arm64", etc.

However, in Ubuntu kernel, some of the flavour names are shared across
multiple architectures, like "generic", "aws", "gcp", etc.

If this information is desired, the substvars feature could be used to
differentiate the Description field in binary package control files
(DEBIAN/control), but I do not believe it is worth the effort.

Signed-off-by: Masahiro Yamada <masahiro.yamada@canonical.com>
Signed-off-by: Paolo Pisati <paolo.pisati@canonical.com>