arm64: dts: qcom: x1e80100-vivobook-s15: mark l12b and l15b always-on
The l12b and l15b supplies are used by components that are not (fully)
described (and some never will be) and must never be disabled.

Mark the regulators as always-on to prevent them from being disabled,
for example, when consumers probe defer or suspend.

adapted-from 20250227081357.25971-1-johan+linaro@kernel.org

Signed-off-by: Jens Glathe <jens.glathe@oldschoolsolutions.biz>