arm64: dts: qcom: x1e001de-devkit: Add HDMI port
Some early units of the devkit have an HDMI port with a DP to HDMI bridge
connected to the &mdss_dp3 instance that normally drives the internal
display. The HDMI port was removed in later revisions, but even they still
have the standard 40-pin eDP header, so it is likely that any (e)DP to HDMI
bridge could be used to add an HDMI port there.

There is no way to tell if there is nothing, the HDMI bridge or even an eDP
panel connected, so let's assume the original intended use case and
describe the HDMI port in the devkit DT. This covers the two most likely
use cases:

 - The HPD line is pulled low on the mainboard side, so if the eDP cable or
   HDMI port is missing the port will remain disconnected.

 - If the eDP cable and HDMI port is present it will detect monitors as
   expected.

If someone wants to deviate from this and connect an actual eDP panel, they
would need to modify the device tree.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>