arm64: dts: qcom: x1e001de-devkit: Add DS1308 RTC
Unlike most X1E laptops, the devkit has a dedicated I2C DS1308 RTC
connected to a small battery. The PMIC RTC could still be used for alarm
wakeup while the board is powered and in suspend, but we describe the PMIC
RTC with "qcom,no-alarm", so just disable it entirely for now.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>