arm64: dts: qcom: x1-crd: Add missing pinctrl for eDP HPD signal
At the moment, we indirectly rely on the boot firmware to set up the
pinctrl for the eDP HPD line coming from the internal display. If the boot
firmware does not configure the display (e.g. because you chose a different
display output in the UEFI settings), then the display fails to come up and
there are several errors in the kernel log:

 [drm:dpu_encoder_phys_vid_wait_for_commit_done:544] [dpu error]vblank timeout: 80020041
 [drm:dpu_kms_wait_for_commit_done:524] [dpu error]wait for commit done returned -110
 [drm:dpu_encoder_frame_done_timeout:2715] [dpu error]enc40 frame done timeout
 ...

Fix this by adding the missing pinctrl for gpio119. The setup seems to be
identical on most boards, so add it to the SoC dtsi next to the other
special GPIO functions.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>