arm64: dts: qcom: x1e001de-devkit: Add speaker amplifiers
Like the X1E CRD, the devkit has 4x WSA8845 on the mainboard that are
available on a pin header on the mainboard. There is nothing connected
there by default, but we should describe it anyway to ensure the chips
get properly shut down (and in case someone would like to use them).

Copy the definitions as-is from x1-crd.dtsi, since the setup is exactly
the same.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>