arm64: dts: qcom: x1e001de-devkit: Add DMICs
The devkit has 4 DMICs on the front side (where the power button is
located). The design is largely identical to the X1E CRD, so copy the
definitions from x1-crd.dtsi to make them work.

The only difference is that the DMICs in the devkit do not appear to work
correctly when driven with 4.8 MHz sample rate, there is lots of noise
coming from one of the channels. 2.4 MHz and 3.2 MHz both seem to work
fine. Use 2.4 MHz, which is also what Windows uses. Note that even the CRD
uses 2.4 MHz on Windows, so we might also run into other X1E devices with
this problem in the future.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>