arm64: dts: qcom: x1e001de-devkit: Add PCIe3 x8 slot
Some versions of the devkit have a full x16 PCIe slot mounted on the
mainboard, connected to the x8 PCIe3 instance of the SoC. Describe it in
the device tree together with the associated regulator. Unlike QCP, there
seems to be just a single controllable 3.3V supply, the 12V supply appears
to be always on.

Not all devkits have the PCIe slot mounted. If it is not present, all the
regulator supplies, signals etc are still present. It will behave just like
when there is no card inserted in the slot. We should still describe it in
the device tree to make sure it can be powered down properly.

Signed-off-by: Stephan Gerhold <stephan.gerhold@linaro.org>