ASoC: qcom: x1e80100: Add Dell XPS13 9345 support
On Dell XPS13 9345, the PCM channels are mapped starting with right
hand side instead of left. So in order to support this, we need to
hardcode the mapping and tie it up to a dedicated board compatible.

So define a match data that brings the mapping as well, for the XPS 13,
while the rest of the boards will fallback to use the mapping based
on number of channels.

Link: https://patch.msgid.link/20251021-dell-xps13-9345-enable-audio-v1-2-6f3f6bbd977b@linaro.org
Signed-off-by: Abel Vesa <abel.vesa@oss.qualcomm.com>